2025Aug07 v1.19.1 1. Rename folders for following MVC naming rules.
