+++ pom.xml
@@ -12,7 +12,7 @@
 
     <groupId>com.GDTW</groupId>
     <artifactId>GDTW_Web</artifactId>
-    <version>1.19.0</version>
+    <version>1.19.1</version>
     <name>GDTW_Web</name>
     <description>Free Web Service 'GDTW.org'</description>
