+++ src/main/java/com/gdtw/dailystatistic/controller/DailyStatisticRestController.java
@@ -1,8 +1,8 @@
 package com.gdtw.dailystatistic.controller;
 
-import com.gdtw.dailystatistic.dto.ChartDataDTO;
+import com.gdtw.dailystatistic.model.ChartDataDTO;
 import com.gdtw.dailystatistic.service.DailyStatisticService;
-import com.gdtw.dailystatistic.dto.TotalServiceStatisticsDTO;
+import com.gdtw.dailystatistic.model.TotalServiceStatisticsDTO;
 import com.gdtw.general.helper.ratelimiter.RateLimiterHelper;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.ResponseEntity;
