+++ src/main/java/com/gdtw/dailystatistic/model/TotalServiceStatisticsDTO.java
@@ -1,4 +1,4 @@
-package com.gdtw.dailystatistic.dto;
+package com.gdtw.dailystatistic.model;
 
 import java.io.Serializable;
