+++ src/main/java/com/gdtw/general/config/JpaConfig.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @Configuration
 @EnableJpaRepositories(basePackages = {
-        "com.gdtw.dailystatistic.model",
+        "com.gdtw.dailystatistic.repository",
         "com.gdtw.imgshare.model",
-        "com.gdtw.shorturl.model",
-        "com.gdtw.reportviolation.model"
+        "com.gdtw.shorturl.repository",
+        "com.gdtw.reportviolation.repository"
 })
 
 public class JpaConfig {
