+++ src/main/java/com/gdtw/imgshare/controller/ImgShareRestController.java
@@ -4,9 +4,9 @@ import com.gdtw.dailystatistic.service.DailyStatisticService;
 import com.gdtw.general.exception.InsufficientDiskSpaceException;
 import com.gdtw.general.helper.ratelimiter.RateLimiterHelper;
 import com.gdtw.general.util.ImgServiceValidatorUtil;
-import com.gdtw.imgshare.dto.AlbumCreationRequestDTO;
+import com.gdtw.imgshare.model.AlbumCreationRequestDTO;
 import com.gdtw.imgshare.service.ImgShareService;
-import com.gdtw.imgshare.dto.TokenRequestDTO;
+import com.gdtw.imgshare.model.TokenRequestDTO;
 import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
