+++ src/main/java/com/gdtw/imgshare/repository/ShareImgAlbumJpa.java
@@ -1,5 +1,6 @@
-package com.gdtw.imgshare.model;
+package com.gdtw.imgshare.repository;
 
+import com.gdtw.imgshare.model.ShareImgAlbumVO;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
