+++ src/main/java/com/gdtw/imgshare/service/ImgSharePersistenceService.java
@@ -6,12 +6,12 @@ import com.gdtw.general.helper.RedisObjectCacheHelper;
 import com.gdtw.general.util.CodecImgFilenameUtil;
 import com.gdtw.general.util.CodecImgIdUtil;
 import com.gdtw.general.helper.JwtHelper;
-import com.gdtw.imgshare.dto.AlbumCreationRequestDTO;
-import com.gdtw.imgshare.dto.ShareImgAlbumInfoDTO;
-import com.gdtw.imgshare.dto.ShareImgInfoDTO;
-import com.gdtw.imgshare.model.ShareImgAlbumJpa;
+import com.gdtw.imgshare.model.AlbumCreationRequestDTO;
+import com.gdtw.imgshare.model.ShareImgAlbumInfoDTO;
+import com.gdtw.imgshare.model.ShareImgInfoDTO;
+import com.gdtw.imgshare.repository.ShareImgAlbumJpa;
 import com.gdtw.imgshare.model.ShareImgAlbumVO;
-import com.gdtw.imgshare.model.ShareImgJpa;
+import com.gdtw.imgshare.repository.ShareImgJpa;
 import com.gdtw.imgshare.model.ShareImgVO;
 import jakarta.transaction.Transactional;
 import org.slf4j.Logger;
