+++ src/main/java/com/gdtw/reportviolation/controller/ReportViolationRestController.java
@@ -1,7 +1,7 @@
 package com.gdtw.reportviolation.controller;
 
 import com.gdtw.general.helper.ratelimiter.RateLimiterHelper;
-import com.gdtw.reportviolation.dto.ReportRequestDTO;
+import com.gdtw.reportviolation.model.ReportRequestDTO;
 import com.gdtw.reportviolation.service.ReportViolationService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.dao.DataIntegrityViolationException;
