+++ src/main/java/com/gdtw/shorturl/controller/ShortUrlRestController.java
@@ -4,12 +4,12 @@ import com.gdtw.dailystatistic.service.DailyStatisticService;
 import com.gdtw.general.exception.ShortUrlBannedException;
 import com.gdtw.general.exception.ShortUrlNotFoundException;
 import com.gdtw.general.helper.ratelimiter.RateLimiterHelper;
-import com.gdtw.general.service.safebrowsing4.SafeBrowsingV4Service;
+import com.gdtw.general.service.safebrowsing4.service.SafeBrowsingV4Service;
 import com.gdtw.general.util.UrlServiceValidatorUtil;
-import com.gdtw.shorturl.dto.CreateShortUrlRequestDTO;
-import com.gdtw.shorturl.dto.GetOriginalUrlDTO;
-import com.gdtw.shorturl.dto.ReturnCreatedShortUrlDTO;
-import com.gdtw.shorturl.dto.ReturnOriginalUrlDTO;
+import com.gdtw.shorturl.model.CreateShortUrlRequestDTO;
+import com.gdtw.shorturl.model.GetOriginalUrlDTO;
+import com.gdtw.shorturl.model.ReturnCreatedShortUrlDTO;
+import com.gdtw.shorturl.model.ReturnOriginalUrlDTO;
 import com.gdtw.shorturl.service.ShortUrlService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
