+++ src/main/java/com/gdtw/shorturl/model/CreateShortUrlRequestDTO.java
@@ -1,4 +1,4 @@
-package com.gdtw.shorturl.dto;
+package com.gdtw.shorturl.model;
 
 public class CreateShortUrlRequestDTO {
