+++ src/main/java/com/gdtw/shorturl/model/ShortUrlInfoDTO.java
@@ -1,4 +1,4 @@
-package com.gdtw.shorturl.dto;
+package com.gdtw.shorturl.model;
 
 import java.io.Serializable;
